+++ lib/ModelMixins/TableMixin.ts
@@ -242,8 +242,15 @@ function TableMixin<T extends Constructor<Model<TableTraits>>>(Base: T) {
     }
 
     @computed
-    get canZoomTo() {
-      return this.activeTableStyle.latitudeColumn !== undefined;
+    get disableZoomTo() {
+      // Disable zoom if only showing imagery parts  (eg region mapping) and no rectangle is defined
+      if (
+        !this.mapItems.find(m => m instanceof DataSource) &&
+        !isDefined(this.cesiumRectangle)
+      ) {
+        return true;
+      }
+      return super.disableZoomTo;
     }
 
     /**