+++ lib/Models/CartoMapCatalogItem.ts
@@ -109,10 +109,6 @@ export default class CartoMapCatalogItem extends MappableMixin(
     return CartoMapCatalogItem.type;
   }
 
-  get canZoomTo() {
-    return isDefined(this.rectangle);
-  }
-
   @computed get mapItems(): MapItem[] {
     if (isDefined(this.imageryProvider)) {
       return [