+++ lib/Models/WebFeatureServiceCatalogItem.ts
@@ -267,8 +267,6 @@ class WebFeatureServiceCatalogItem extends ExportableMixin(
   ];
 
   static readonly type = "wfs";
-  readonly canZoomTo = true;
-
   @observable
   private geojsonCatalogItem: GeoJsonCatalogItem | undefined;